CORRECT RATE CONSTANTS FOR SAPRC99 MECHANISMS
Corrects erroneous rate constants in the SAPRC99 mechanisms. For several
reactions, the rate constants were replaced with “2.20D-10” (with no
temperature dependence). In addition, there were additional weird reactions
with the same rate constant that are not present in the original SAPRC99
mechanism. This commit replaces the rate constants with the ones from the
original SAPRC99 publication (Carter, 2000) and removes the additional
reactions.  Reference: Carter et al., 2000: Documentation of the SAPRC-99
chemical mechanism for VOC reactivity assessment