SYNCHRONIZE THE ADAPTIVE TIME STEP WITH RADT
In dyn_em/adapt_timestep_em.F, forces the adaptive time step if used to
always fall exactly on the expected radiation call time. This was not
done in base WRF, and as a result the radiation code was not called every
radt as it expects. This seems to have caused issues in the calculation
of the cosine zenith angle and in the total accumulated irradiance, and
could cause biases of several K for 2-m Temperature.